+++ Digital Pantry/Model/Recipe.swift
@@ -9,13 +9,13 @@ import Foundation
 
 struct Recipe {
     //properties/members of Recipe
-    var recipeID: Int
+    var recipeID: Int64
     var recipeName: String
     var recipeDescription: String
-    var cookingTime: Int
-    var complexity: Int
+    var cookingTime: Int64
+    var complexity: Int64
     //initialised where we have not obtained a UniqueIdentifier (recipeID) as yet
-    init?(foodCategoryID: Int, recipeName: String, recipeDescription: String, cookingTime: Int, complexity: Int) {
+    init?(foodCategoryID: Int64, recipeName: String, recipeDescription: String, cookingTime: Int64, complexity: Int64) {
         self.recipeID = -1
         self.recipeName = recipeName
         self.recipeDescription = recipeDescription
@@ -24,7 +24,7 @@ struct Recipe {
     }
     
     //use case: calling data back from DB/table
-    init?(recipeID: Int, foodCategoryID: Int, recipeName: String, recipeDescription: String, cookingTime: Int, complexity: Int) {
+    init?(recipeID: Int64, foodCategoryID: Int64, recipeName: String, recipeDescription: String, cookingTime: Int64, complexity: Int64) {
         self.recipeID = recipeID
         self.recipeName = recipeName
         self.recipeDescription = recipeDescription
@@ -32,11 +32,11 @@ struct Recipe {
         self.complexity = complexity
     }
     
-    mutating func setRecipeID(recipeID: Int) { //once the DB has a recipeID, use this to set it
+    mutating func setRecipeID(recipeID: Int64) { //once the DB has a recipeID, use this to set it
         self.recipeID = recipeID
     }
     
-    func getRecipeID() -> Int { //retrieve recipeID
+    func getRecipeID() -> Int64 { //retrieve recipeID
         return recipeID
     }
     
@@ -56,19 +56,19 @@ struct Recipe {
         self.recipeDescription = recipeDescription
     }
     
-    mutating func setCookingTime(cookingTime: Int) { //set cookingTime
+    mutating func setCookingTime(cookingTime: Int64) { //set cookingTime
         self.cookingTime = cookingTime
     }
     
-    func getCookingTime() -> Int { //retrieve cookingTime
+    func getCookingTime() -> Int64 { //retrieve cookingTime
         return cookingTime
     }
     
-    mutating func setComplexity(complexity: Int) { //set complexity
+    mutating func setComplexity(complexity: Int64) { //set complexity
         self.complexity = complexity
     }
     
-    func getComplexity() -> Int { //retrieve complexity
+    func getComplexity() -> Int64 { //retrieve complexity
         return complexity
     }
 }
